% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{chroma_connect}
\alias{chroma_connect}
\title{ChromaDB Client}
\usage{
chroma_connect(
  host = "http://localhost",
  port = 8000L,
  api_path = "/api/v2",
  verify = TRUE
)
}
\arguments{
\item{host}{The host URL of the ChromaDB server. Default is "http://localhost".}

\item{port}{The port number of the ChromaDB server. Default is 8000.}

\item{api_path}{The API path. Default is "/api/v2".}

\item{verify}{Whether to verify the connection. Default is TRUE.}
}
\value{
A ChromaDB client object
}
\description{
Create a new ChromaDB client connection
}
