% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{chroma_docker_running}
\alias{chroma_docker_running}
\title{Check ChromaDB Docker Container Status}
\usage{
chroma_docker_running(container_name = "chromadb")
}
\arguments{
\item{container_name}{A string specifying the name of the Docker container to check.}
}
\value{
TRUE if container is running and FALSE otherwise.
}
\description{
This function checks the status of the ChromaDB Docker container.
}
