% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{chroma_docker_stop}
\alias{chroma_docker_stop}
\title{Stop ChromaDB Docker Container}
\usage{
chroma_docker_stop(container_name = "chromadb")
}
\arguments{
\item{container_name}{A string specifying the name of the Docker container to stop.}
}
\value{
Invisibly returns \code{TRUE} if the container was stopped or is not running.
}
\description{
This function stops the running ChromaDB Docker container.
It uses the \code{processx} package to issue the Docker stop command.
}
