% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{count_collections}
\alias{count_collections}
\title{Count Collections in a Database}
\usage{
count_collections(
  client,
  tenant = "default_tenant",
  database = "default_database"
)
}
\arguments{
\item{client}{A ChromaDB client object}

\item{tenant}{The tenant name (default: "default")}

\item{database}{The database name (default: "default")}
}
\value{
Number of collections in the database
}
\description{
Count Collections in a Database
}
