% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{create_database}
\alias{create_database}
\title{Create a Database}
\usage{
create_database(client, name, tenant = "default_tenant")
}
\arguments{
\item{client}{A ChromaDB client object}

\item{name}{The name of the database}

\item{tenant}{The tenant name}
}
\value{
NULL invisibly on success
}
\description{
Create a Database
}
