% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenants.R
\name{create_tenant}
\alias{create_tenant}
\title{Create a Tenant}
\usage{
create_tenant(client, name)
}
\arguments{
\item{client}{A ChromaDB client object}

\item{name}{The name of the tenant}
}
\value{
A tenant object containing the tenant details
}
\description{
Create a Tenant
}
