% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{delete_collection}
\alias{delete_collection}
\title{Delete a Collection}
\usage{
delete_collection(
  client,
  name,
  tenant = "default_tenant",
  database = "default_database"
)
}
\arguments{
\item{client}{A ChromaDB client object}

\item{name}{The name of the collection}

\item{tenant}{The tenant name (default: "default")}

\item{database}{The database name (default: "default")}
}
\value{
Invisible NULL on success
}
\description{
Delete a Collection
}
