% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\name{delete_documents}
\alias{delete_documents}
\title{Delete Documents from a Collection}
\usage{
delete_documents(
  client,
  collection_name,
  ids = NULL,
  where = NULL,
  tenant = "default_tenant",
  database = "default_database"
)
}
\arguments{
\item{client}{A ChromaDB client object}

\item{collection_name}{Name of the collection}

\item{ids}{Vector of document IDs to delete}

\item{where}{Optional filtering conditions}

\item{tenant}{The tenant name (default: "default")}

\item{database}{The database name (default: "default")}
}
\value{
NULL invisibly on success
}
\description{
Delete Documents from a Collection
}
