% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{get_auth_identity}
\alias{get_auth_identity}
\title{Get Authentication Identity}
\usage{
get_auth_identity(client)
}
\arguments{
\item{client}{A ChromaDB client object}
}
\value{
Authentication identity information
}
\description{
Get Authentication Identity
}
