% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{get_collection}
\alias{get_collection}
\title{Get a Collection}
\usage{
get_collection(
  client,
  name,
  tenant = "default_tenant",
  database = "default_database"
)
}
\arguments{
\item{client}{A ChromaDB client object}

\item{name}{The name of the collection}

\item{tenant}{The tenant name (default: "default")}

\item{database}{The database name (default: "default")}
}
\value{
A collection object
}
\description{
Get a Collection
}
