% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_database}
\alias{get_database}
\title{Get a Database}
\usage{
get_database(client, name, tenant = "default_tenant")
}
\arguments{
\item{client}{A ChromaDB client object}

\item{name}{The name of the database}

\item{tenant}{The tenant name}
}
\value{
NULL invisibly on success
}
\description{
Get a Database
}
