% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{get_server_info}
\alias{get_server_info}
\title{Get ChromaDB Server Information}
\usage{
get_server_info(client)
}
\arguments{
\item{client}{A ChromaDB client object}
}
\value{
List containing server information
}
\description{
Returns server capabilities and settings.
}
