% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenants.R
\name{get_tenant}
\alias{get_tenant}
\title{Get a Tenant}
\usage{
get_tenant(client, name)
}
\arguments{
\item{client}{A ChromaDB client object}

\item{name}{The name of the tenant}
}
\value{
A tenant object containing the tenant details
}
\description{
Get a Tenant
}
