% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{get_version}
\alias{get_version}
\title{Get ChromaDB Server Version}
\usage{
get_version(client)
}
\arguments{
\item{client}{A ChromaDB client object}
}
\value{
Server version string
}
\description{
Get ChromaDB Server Version
}
