% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{heartbeat}
\alias{heartbeat}
\title{Check ChromaDB Server Heartbeat}
\usage{
heartbeat(client)
}
\arguments{
\item{client}{A ChromaDB client object}
}
\value{
Server heartbeat response as a numeric value
}
\description{
Check ChromaDB Server Heartbeat
}
