% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{list_collections}
\alias{list_collections}
\title{List Collections in a Database}
\usage{
list_collections(
  client,
  tenant = "default_tenant",
  database = "default_database",
  limit = NULL,
  offset = NULL
)
}
\arguments{
\item{client}{A ChromaDB client object}

\item{tenant}{The tenant name (default: "default")}

\item{database}{The database name (default: "default")}

\item{limit}{Maximum number of collections to return (optional)}

\item{offset}{Number of collections to skip (optional)}
}
\value{
List of collections
}
\description{
List Collections in a Database
}
