% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{pre_flight_checks}
\alias{pre_flight_checks}
\title{Get ChromaDB Server Information}
\usage{
pre_flight_checks(client)
}
\arguments{
\item{client}{A ChromaDB client object}
}
\value{
List containing server information
}
\description{
Returns server capabilities and settings.
}
