% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{reset}
\alias{reset}
\title{Reset ChromaDB}
\usage{
reset(client)
}
\arguments{
\item{client}{A ChromaDB client object}
}
\value{
TRUE on success
}
\description{
This function resets the entire ChromaDB instance. Use with caution as this will delete all data.
Note: This function requires setting ALLOW_RESET=TRUE in the environment variables
or allow_reset=True in the ChromaDB Settings.
}
