% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\name{update_documents}
\alias{update_documents}
\title{Update Documents in a Collection}
\usage{
update_documents(
  client,
  collection_name,
  ids,
  documents = NULL,
  metadatas = NULL,
  embeddings = NULL,
  tenant = "default_tenant",
  database = "default_database"
)
}
\arguments{
\item{client}{A ChromaDB client object}

\item{collection_name}{Name of the collection}

\item{ids}{Vector of document IDs to update}

\item{documents}{List of new document contents}

\item{metadatas}{List of new metadata}

\item{embeddings}{Optional new pre-computed embeddings}

\item{tenant}{The tenant name (default: "default")}

\item{database}{The database name (default: "default")}
}
\value{
NULL invisibly on success
}
\description{
Update Documents in a Collection
}
