% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\name{upsert_documents}
\alias{upsert_documents}
\title{Upsert Documents to a Collection}
\usage{
upsert_documents(
  client,
  collection_name,
  documents,
  metadatas = NULL,
  ids = NULL,
  embeddings = NULL,
  uris = NULL
)
}
\arguments{
\item{client}{A ChromaDB client object}

\item{collection_name}{Name of the collection}

\item{documents}{List of documents to upsert}

\item{metadatas}{List of metadata for each document}

\item{ids}{Vector of unique IDs for the documents}

\item{embeddings}{Optional pre-computed embeddings}

\item{uris}{Optional vector of URIs associated with the documents}
}
\value{
Response from the API
}
\description{
Upsert Documents to a Collection
}
